Renamed the appbar component
